# dj-connect-back/storage/api-docs/api-docs.yaml
openapi: 3.0.0
info:
  title: 'DJ Connect API'
  description: 'API documentation for DJ Connect. All requests must include header Telegram-Init-Data: ${TWA.initData()}'
  contact:
    name: 'Contact Developer'
    url: 'https://t.me/beilec'
  license:
    name: 'Contact Lead'
    url: 'https://t.me/alievdenis1'
  version: 1.0.0
servers:
  -
    url: 'http://localhost:8082/api/v1'
    description: 'DJ Connect API Server'
  -
    url: 'https://dj-connect.xyz/api/v1'
paths:
...
  /orders:
    post:
      tags:
        - Order
      summary: 'Create an order'
      description: 'Allows a user to create an order for a track'
      operationId: createOrder
      requestBody:
        required: true
        content:
          application/json:
            schema:
              properties:
                dj_id:
                  type: integer
                  example: 1
                track_id:
                  type: integer
                  example: 1
                price:
                  type: number
                  format: float
                  example: 19.99
                message:
                  type: string
                  example: 'Please play this track!'
              type: object
      responses:
        200:
          description: 'Successful operation'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        400:
          description: 'Bad Request'
        401:
          description: Unauthorized
        404:
          description: 'DJ or Track not found'
      security:
        -
          telegramAuth: []
  '/orders/{order_id}/accept':
    patch:
      tags:
        - Order
      summary: 'Accept an order'
      description: 'Allows a DJ to accept an order'
      operationId: acceptOrder
      parameters:
        -
          name: order_id
          in: path
          description: 'Order ID'
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              properties:
                price:
                  type: number
                  format: float
                  example: 19.99
                message:
                  type: string
                  example: 'Order accepted'
              type: object
      responses:
        200:
          description: 'Order accepted successfully'
          content:
            application/json:
              schema:
                properties:
                  order:
                    $ref: '#/components/schemas/Order'
                  transaction:
                    $ref: '#/components/schemas/Transaction'
                type: object
      security:
        -
          telegramAuth: []
  '/orders/{order_id}/decline':
    patch:
      tags:
        - Order
      summary: 'Decline an order'
      description: 'Allows a DJ to decline an order'
      operationId: declineOrder
      parameters:
        -
          name: order_id
          in: path
          description: 'Order ID'
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              properties:
                message:
                  type: string
                  example: 'Order declined'
              type: object
      responses:
        200:
          description: 'Order declined successfully'
          content:
            application/json:
              schema:
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: 'Order declined'
                type: object
        400:
          description: 'Bad Request'
        401:
          description: Unauthorized
        404:
          description: 'Order not found'
      security:
        -
          telegramAuth: []
  '/orders/{order_id}/cancel':
    patch:
      tags:
        - Order
      summary: 'Cancel an order'
      description: 'Allows a user to cancel an order and associated transactions'
      operationId: cancelOrder
      parameters:
        -
          name: order_id
          in: path
          description: 'Order ID'
          required: true
          schema:
            type: integer
      responses:
        200:
          description: 'Order and associated transactions cancelled successfully'
          content:
            application/json:
              schema:
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: 'Order and associated transactions cancelled'
                type: object
        400:
          description: 'Bad Request'
        401:
          description: Unauthorized
        404:
          description: 'Order not found'
      security:
        -
          telegramAuth: []
  '/dj/{dj_id}/orders':
    get:
      tags:
        - Order
      summary: 'Get orders for DJ'
      description: 'Returns the list of orders for a specific DJ'
      operationId: getOrdersForDJ
      parameters:
        -
          name: dj_id
          in: path
          description: 'DJ ID'
          required: true
          schema:
            type: integer
      responses:
        200:
          description: 'Successful operation'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        400:
          description: 'Bad Request'
        401:
          description: Unauthorized
        404:
          description: 'DJ not found'
      security:
        -
          telegramAuth: []
  /user/orders:
    get:
      tags:
        - Order
      summary: 'Get orders for user'
      description: 'Returns the list of orders for the authenticated user'
      operationId: getOrdersForUser
      responses:
        200:
          description: 'Successful operation'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        400:
          description: 'Bad Request'
        401:
          description: Unauthorized
      security:
        -
          telegramAuth: []
...
components:
  schemas:
    Order:
      title: Order
      required:
        - user_id
        - dj_id
        - track_id
        - price
        - status
      properties:
        id:
          description: 'Order ID'
          type: integer
          example: 1
        user_id:
          description: 'User ID'
          type: integer
          example: 1
        dj_id:
          description: 'DJ ID'
          type: integer
          example: 1
        track_id:
          description: 'Track ID'
          type: integer
          example: 1
        price:
          description: 'Price of the order'
          type: number
          format: float
          example: 19.99
        message:
          description: 'Message for the order'
          type: string
          example: 'Please play this track!'
        status:
          description: 'Status of the order'
          type: string
          enum:
            - pending
            - accepted
            - declined
            - price_changed
          example: pending
        created_at:
          description: 'Creation timestamp'
          type: string
          format: date-time
          example: '2024-07-01T00:00:00Z'
        updated_at:
          description: 'Last update timestamp'
          type: string
          format: date-time
          example: '2024-07-01T00:00:00Z'
        is_paid:
          description: 'Is the order paid based on the transaction status'
          type: boolean
          example: 'false'
        transactions:
          description: 'Array of transaction details'
          type: array
          items:
            $ref: '#/components/schemas/Transaction'
      type: object
    Transaction:
      title: Transaction
      required:
        - order_id
        - amount
        - status
        - payment_url
      properties:
        id:
          description: 'Transaction ID'
          type: integer
          example: 1
        order_id:
          description: 'Order ID'
          type: integer
          example: 1
        amount:
          description: 'Transaction amount'
          type: number
          format: float
          example: 19.99
        payment_url:
          description: 'Payment URL'
          type: string
          example: 'https://payment.gateway.com/pay?order_id=1'
        status:
          description: 'Transaction status'
          type: string
          enum:
            - pending
            - paid
            - cancelled
          example: pending
        created_at:
          description: 'Creation timestamp'
          type: string
          format: date-time
          example: '2024-07-01T00:00:00Z'
        updated_at:
          description: 'Last update timestamp'
          type: string
          format: date-time
          example: '2024-07-01T00:00:00Z'
      type: object
...
  securitySchemes:
    telegramAuth:
      type: apiKey
      in: header
      name: Telegram-Init-Data
      description: 'Value of TWA.initData() from Telegram Web App.'
...
security:
  -
    telegramAuth: []